==> riboseq/04_riboseq_p_sites_diff_expr_analysis.Rmd <==
---
title: "Identifying ORFs differentially translated between MYC high and MYC low samples (cell lines only)"
output: html_notebook
---

# Description
This R Notebook is used to analyze differentially translated open reading frames (ORFs) in MYC high and MYC low cell lines. The analysis includes loading metadata, filtering samples, running DESeq2, creating volcano plots, and generating PCA plots.

Inputs:
- Metadata file: 01.26.2023_Table_S1.xlsx
- ORF reference table: all_ORFs_ref_table.csv (Based on Table S1)
- Riboseq P-sites counts: S1_riboseq_P_sites_counts.csv (Table S1K)

## Output files:
1. ORF level differential translation analysis results: ORF_level_diff_transl_analysis_results.csv

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

suppressPackageStartupMessages({
  library(tidyverse)
  library(RColorBrewer)
  library(DESeq2)
  library(ggrepel)
  library(EnhancedVolcano)
  library(ggbeeswarm)
})
```

# Set file locations

```{r file-locations}
# Metadata file
metadata_location <- "Table_S1.xlsx"

# ORF reference table
ref_table_location <- "/annotation_files/all_ORFs_ref_table.csv"
```

# Load and filter metadata

```{r load-metadata}
# Load metadata table S1A
sample_metadata <- readxl::read_xlsx(metadata_location, 
  sheet = 1, 
  col_names = T, 
  skip = 2, 
  .name_repair = "universal")

# Filter samples with both RNAseq and riboseq data
sample_metadata_complete <- sample_metadata[which(
  !is.na(sample_metadata$RNASeq_file_names) &
    !is.na(sample_metadata$RiboSeq_file_names..final.)), ]

sample_metadata_filtered <- subset(sample_metadata_complete, Sample_Type == "Cell line")
```

# Load data and match P-site counts columns

```{r load-data}
counts_psites <- read.csv("/riboseq/results/tables/S1_riboseq_P_sites_counts.csv", row.names = 1, check.names = F)
ref_table <- read.csv(ref_table_location)
```

```{r match-columns}
# Extract the column names from rna_counts that match the RNASeq_file_names in sample_metadata_celllines
match_counts <- unique(str_extract(sample_metadata_filtered$Sample_ID, str_c(colnames(counts_psites), collapse = "|")))

# Reorder the columns of P-site counts data frame to match the order of metadata
counts_psites_ordered <- counts_psites[, match_counts]

# Find matching metadata
metadata_matches <- grep(paste(match_counts, collapse = "|"), sample_metadata_filtered$Sample_ID)
sample_metadata_matched <- sample_metadata_filtered[metadata_matches, ]

colnames(counts_psites) <- sample_metadata_filtered$Sample_ID

# Fix metadata
coldata_select <- sample_metadata_filtered
colnames(coldata_select) <- gsub(" ", "_", colnames(coldata_select))
colnames(coldata_select) <- gsub("-", "_", colnames(coldata_select))
```


# Differential translation analysis (Table S1O)

```{r}
# Run DESeq2 to analyze differential translation
dds <- DESeqDataSetFromMatrix(countData = counts_psites_ordered,
                              colData = sample_metadata_matched,
                              design = ~ MYC.group)

# Run DESeq2
dds <- DESeq(dds)

# Transform DESeq2 output
vsd <- DESeq2::vst(dds, blind = F)

# Store dds counts and vsd counts in matrices
mat_dds <- data.frame(counts(dds, normalized = T))
mat_vsd <- assay(vsd)

# Get DESeq2 results
res <- results(dds, contrast = c("MYC.group", "MYC high", "MYC low"), alpha = 0.01)

# Make readable data frame
res_df <- data.frame(res) %>%
  rownames_to_column("orf_id") %>%
  left_join(ref_table) %>%
  mutate(orf_name = paste0(.$gene_name, "_", .$biotype))

res_df[which(is.na(res_df$gene_name)), "orf_name"] <- res_df[which(is.na(res_df$gene_name)), "orf_id"]

# Save ORF level differential translation analysis results
write.csv(res_df, "ORF_level_diff_transl_analysis_results.csv")
```

# Visualization: Volcano Plot (Fig 1K)

```{r volcano-plot, eval = FALSE}
# Create a volcano plot for the differential translation analysis results of non-canonical ORFs

# Select non-canonical ORF results
res_NC_df <- subset(res_df, canonical == "Non-canonical")
res_2 <- res_NC_df

# Select top 5 downregulated ORFs (log2FC < -2, based on padj)
top5down <- res_NC_df %>%
  subset(log2FoldChange < 2) %>%
  arrange(.$padj) %>%
  head(5)

# Select top 5 upregulated ORFs (log2FC < 2, based on padj)
top5up <- res_NC_df %>%
  subset(log2FoldChange > 2) %>%
  arrange(.$padj) %>%
  head(5) 

# Combine top 5 up and down ORFs
top <- rbind(top5down, top5up)

# Remove trailing '_biotype' part from ORF names to label
top$orf_name <- gsub("_.*$", "", top$orf_name)

# Remove the top 10 differentially regulated ORFs from results dataset
res_2 <- res_2[-which(res_2$orf_id %in% top$orf_id), ]

# Set ORF name of the rest of the ORFs to NA (so they won't be labeled)
res_2$orf_name <- NA

# Add results of the top 10 ORFs, including their ORF names (so only these will be labeled)
res_2 <- rbind(res_2, top)

# Make custom keyvals table
keyvals <- data.frame(res_2) %>%
  select(c("orf_id", "biotype", "padj"))

cols_biotype <- brewer.pal(6, "Set1")

# Set color of points by ORF biotype
keyvals[which(keyvals$padj < 0.01 & keyvals$biotype == "lncRNA"), "color"] <- cols_biotype[1]
keyvals[which(keyvals$padj < 0.01 & keyvals$biotype == "uORF"), "color"] <- cols_biotype[2]
keyvals[which(keyvals$padj < 0.01 & keyvals$biotype == "uoORF"), "color"] <- cols_biotype[3]
keyvals[which(keyvals$padj < 0.01 & keyvals$biotype == "dORF"), "color"] <- cols_biotype[4]
keyvals[which(keyvals$padj < 0.01 & keyvals$biotype == "intORF"), "color"] <- cols_biotype[5]
keyvals[which(keyvals$padj < 0.01 & keyvals$biotype == "doORF"), "color"] <- cols_biotype[6]
keyvals$name <- keyvals$biotype
keyvals[which(is.na(keyvals$color)), "name"] <- NA
keyvals[which(is.na(keyvals$color)), "color"] <- "grey"

# Make named list from keyvals table (needed for volcanoplot)
keyvals_list <- keyvals$color
names(keyvals_list) <- keyvals$name

# Plot the volcanoplot
plt <- EnhancedVolcano(
  res_2,
  lab = res_2$orf_name, legendLabels = c("NS", "NS", "NS", "p < 0.01 & log2FC > 2 | < -2"),
  colCustom = keyvals_list,

  x = "log2FoldChange", 
  y = "padj", 
  title = NULL,
  subtitle = NULL, colAlpha = 1, col = c("grey", "grey", "grey", "red"), pCutoff = 0.01, FCcutoff = 0, xlim = c(-12, 12), labSize = 3, labFace = "bold", max.overlaps = 10, drawConnectors = T, pointSize = 1, lengthConnectors = unit(0, "mm"), axisLabSize = 15, legendLabSize = 10)
plt

```

# Visualization: PCA Plot

```{r pca-plot, eval = FALSE}
# Create a PCA plot for the P-site counts data
pcaData <- plotPCA(vsd, intgroup = c("MYC.group"), returnData = T) 
  
percentVar <- round(100 * attr(pcaData, "percentVar"))

ggplot(pcaData, aes(PC1, PC2, color=MYC.group)) +
  scale_color_brewer(palette = "Set1") +
  geom_point(size=3) +
  xlab(paste0("PC1: ",percentVar[1],"% variance")) +
  ylab(paste0("PC2: ",percentVar[2],"% variance")) + theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
                                                           panel.background = element_blank(), axis.line = element_line(colour = "black"))+ theme(aspect.ratio=1) +
  labs(color = "MYC level", title = "P-site counts (cell lines)")
```

# Visualization: plot ORF couns

## ASNSD1-uORF normalized counts plot (Fig 4A)

```{r ASDURF-counts}
# Create a dotplot of normalized ASNSD1-uORF P site counts in MYC high and MYC low samples

ASDURF_counts <- plotCounts(dds = dds, gene = "ENST00000607829_CDS", intgroup = "MYC.group", returnData = T, normalized = T)

ASDURF_plt <- ggplot(data = ASDURF_counts, aes(y = count, x = MYC.group, color = MYC.group)) +
  scale_color_brewer(palette = "Set1") +
  labs(title = "ASDURF P-site counts", x = NULL, y = "P-site counts") +
  geom_beeswarm() +
  theme_classic() +
  ggpubr::stat_compare_means(comparisons = list(c("MYC high", "MYC low")), label = "p.format", ref.group = "MYC low", label.y = max(ASDURF_counts$count)* 1.1)
ASDURF_plt

# Save the plot
ggsave(plot = ASDURF_plt,
       filename = "4A_ASDURF_Riboseq.pdf", device = "pdf", path = , width = unit(6, "cm"), height = unit(6, "cm"))

```

## Other ORFs

```{r}
plot_ORF_counts <- function(dds, ref_table, gene_string) {
  # Get a vector of all ORF IDs that belong to the gene of interest
  orf_ids <- ref_table$orf_id[grepl(gene_string, ref_table$gene_name)]

  # Initialize an empty list to store the plots
  plot_list <- list()

  # Loop over all ORF IDs and generate a plot for each
  for (orf_id in orf_ids) {
    # Get gene_name and biotype for the current ORF
    gene_name <- ref_table$gene_name[ref_table$orf_id == orf_id]
    biotype <- ref_table$biotype[ref_table$orf_id == orf_id]
    
    # Generate plot
    cts <- plotCounts(dds = dds, gene = orf_id, intgroup = "MYC.group", returnData = T, normalized = T)
    
    plt <- ggplot(data = cts, aes(y = count, x = MYC.group, color = MYC.group)) +
      scale_color_brewer(palette = "Set1") +
      labs(title = paste0(c(gene_name, "_", biotype, " P-site counts"), collapse = ""), 
           subtitle = paste0("ORF ID: ", orf_id),
           x = NULL, y = "P-site counts") +
      geom_beeswarm() +
      theme_classic() +
      ggpubr::stat_compare_means(comparisons = list(c("MYC high", "MYC low")), label = "p.format", ref.group = "MYC low", label.y = max(cts$count)* 1.1)
    
    # Add plot to the list
    plot_list[[orf_id]] <- plt
  }

  return(plot_list)
}

```

### TPT1-AS1_lncRNA

```{r}
# pdf("/riboseq/results/plots/TPT1_AS1.pdf")
plot_ORF_counts(dds, ref_table, "TPT1-AS1")
# dev.off()
```


### ZNF525_lncRNA

```{r}
plot_ORF_counts(dds, ref_table, "ZNF525")
```


### LOC101927027_lncRNA
Also known as CHROMR

```{r}
# pdf("riboseq/results/plots/CHROMR.pdf")
plot_ORF_counts(dds, ref_table, "CHROMR")
# dev.off()
```

### PAX8-AS1_lncRNA

```{r}
plot_ORF_counts(dds, ref_table, "PAX8-AS1")
```

### RP1-261D10.2.1_lncRNA 
Previously known asAC005476.2, which is how it was annotated in ref_table. Now known as SIPA1L1-AS1.

```{r}
plot_ORF_counts(dds, ref_table, "AC005476.2")
```

### G049876_lncRNA -> what is this?


### LINC01399_lncRNA
Not in data

### LINC00888_lncRNA
```{r, fig.width=4, fig.height=4}
# pdf("riboseq/results/plots/LINC00888.pdf")
plot_ORF_counts(dds, ref_table, "LINC00888")
# dev.off()
```

### LOC401320.1_lncRNA
Now known as GARS1-DT

```{r}
plot_ORF_counts(dds, ref_table, "GARS1-DT")
```


### RP11-262H14.3_lncRNA
Now known as LERFS

```{r}
plot_ORF_counts(dds, ref_table, "LERFS")
```
### PFD(L) complex members
```{r}
plot_ORF_counts(dds, ref_table, "PDRG1")
```

```{r}
plot_ORF_counts(dds, ref_table, "UXT")
```
```{r}
plot_ORF_counts(dds, ref_table, "URI1")
```

```{r}
plot_ORF_counts(dds, ref_table, "PFDN4")
```
```{r}
plot_ORF_counts(dds, ref_table, "PFDN2")
```
```{r}
plot_ORF_counts(dds, ref_table, "PFDN6")
```

```{r}
plot_ORF_counts(dds, ref_table, "ASDURF")
```
```{r}
plot_ORF_counts(dds, ref_table, "ASNSD1")
```

```{r}
plot_ORF_counts(dds, ref_table, "PFDN5")
```

```{r}
plot_ORF_counts(dds, ref_table, "VBP1")
```

```{r}
plot_ORF_counts <- function(dds, ref_table, genes) {
  # Initialize a data frame to collect results
  results_df <- data.frame()

  # Get the results table from the DESeq object
  deseq_results <- results(dds)

  # Loop over all genes
  for (gene_string in genes) {
    # Get a vector of all ORF IDs that belong to the gene of interest
    orf_ids <- ref_table$orf_id[grepl(gene_string, ref_table$gene_name)]

    # Temporary data frame to store results for this gene
    gene_results <- data.frame()

    # Loop over all ORF IDs and gather results
    for (orf_id in orf_ids) {
      # Get gene_name and biotype for the current ORF
      gene_name <- ref_table$gene_name[ref_table$orf_id == orf_id]
      biotype <- ref_table$biotype[ref_table$orf_id == orf_id]
    
      # Generate plot
      cts <- plotCounts(dds = dds, gene = orf_id, intgroup = "MYC.group", returnData = T, normalized = T)
    
      # Get P value from the DESeq results table
      p_value <- deseq_results[orf_id, "padj"]

      # Create a temporary data frame to store the results
      temp_df <- data.frame(gene_name = gene_name,
                            orf_id = orf_id,
                            MYC_high = mean(cts$count[cts$MYC.group == "MYC high"]),
                            MYC_low = mean(cts$count[cts$MYC.group == "MYC low"]),
                            p_value = p_value)
    
      # Append to gene_results
      gene_results <- rbind(gene_results, temp_df)
    }

    # For this gene, keep only the row with the lowest p_value
    gene_results <- gene_results[which.min(gene_results$p_value),]

    # Append to overall results
    results_df <- rbind(results_df, gene_results)
  }

  # Return the results
  return(results_df)
}

# List of genes to analyze
genes_to_analyze <- c("TPT1-AS1", "ZNF525", "CHROMR", "PAX8-AS1", "AC005476.2", "LINC00888", "GARS1-DT", "LERFS")

# Call the function with the list of genes
all_results <- plot_ORF_counts(dds, ref_table, genes_to_analyze)

# Order by P value
all_results <- all_results[order(all_results$p_value),]


# Plot the combined bar plot
ggplot(all_results, aes(x = reorder(gene_name, p_value), y = MYC_high)) +
  geom_bar(stat = "identity") +
  geom_bar(aes(y = MYC_low), stat = "identity", fill = "red") +
  theme(axis.text.x = element_text(angle = 90, hjust = 1)) +
  labs(x = "Gene Name", y = "P-site counts", title = "Comparison of MYC High vs MYC Low")

```
```{r}
library(tidyverse)

# Reshape the data into a longer format
all_results_long <- all_results %>%
  gather(key = "condition", value = "count", MYC_high, MYC_low)

# Plot the combined bar plot with bars side by side
ggplot(all_results_long, aes(x = reorder(gene_name, p_value), y = count, fill = condition)) +
  geom_bar(stat = "identity", position = "dodge") +
  theme(axis.text.x = element_text(angle = 90, hjust = 1)) +
  labs(x = "Gene Name", y = "P-site counts", title = "Comparison of MYC High vs MYC Low") +
  scale_fill_brewer(palette = "Set1")

```




```{r}
plot_ORF_counts <- function(dds, ref_table, genes) {
  # Initialize a data frame to collect results
  results_df <- data.frame()

  # Get the results table from the DESeq object
  deseq_results <- results(dds)

  # Loop over all genes
  for (gene_string in genes) {
    # Get a vector of all ORF IDs that belong to the gene of interest
    orf_ids <- ref_table$orf_id[grepl(gene_string, ref_table$gene_name)]
    
    # Temporary data frame to store results for this gene
    gene_results <- data.frame()

    # Loop over all ORF IDs and gather results
    for (orf_id in orf_ids) {
      # Get gene_name for the current ORF
      gene_name <- ref_table$gene_name[ref_table$orf_id == orf_id]
      
      # Get P value from the DESeq results table
      p_value <- deseq_results[orf_id, "padj"]

      # Get the counts
      cts <- plotCounts(dds = dds, gene = orf_id, intgroup = "MYC.group", returnData = T, normalized = T)
      MYC_high = mean(cts$count[cts$MYC.group == "MYC high"])
      MYC_low = mean(cts$count[cts$MYC.group == "MYC low"])
      
      # Create a temporary data frame to store the results
      temp_df <- data.frame(gene_name = gene_name,
                            MYC_high = MYC_high,
                            MYC_low = MYC_low,
                            p_value = p_value)
    
      # Append to gene_results
      gene_results <- rbind(gene_results, temp_df)
    }

    # For this gene, keep only the row with the lowest p_value
    gene_results <- gene_results[which.min(gene_results$p_value),]

    # Append to overall results
    results_df <- rbind(results_df, gene_results)
  }

  # Return the results
  return(results_df)
}

```

```{r}
genes_to_analyze <- c("TPT1-AS1", "ZNF525", "CHROMR", "PAX8-AS1", "AC005476.2", "LINC00888", "GARS1-DT", "LERFS")
all_results <- plot_ORF_counts(dds, ref_table, genes_to_analyze)
```

```{r}
# Order by p-value
all_results <- all_results[order(all_results$p_value),]

# Reshape the data into a longer format
all_results_long <- all_results %>%
  select(gene_name, MYC_high, MYC_low) %>%
  gather(key = "condition", value = "count", MYC_high, MYC_low)


write.csv(all_results, "riboseq/results/tables/tophits_Psite_counts_summary.csv")
```

### Step 4: Plot the Bar Plot

```{r}
library(ggplot2)
library(RColorBrewer)

# Plot the combined bar plot with bars side by side
p <- ggplot(all_results_long, aes(x = factor(gene_name, levels = all_results$gene_name), y = count, fill = condition)) +
  geom_bar(stat = "identity", position = "dodge") +
  scale_fill_brewer(palette = "Set1") +
  theme(axis.text.x = element_text(angle = 90, hjust = 1)) +
  labs(x = "Gene Name", y = "P-site counts", title = "Comparison of MYC High vs MYC Low")

# Add p-values as annotations above the bars
for (i in 1:nrow(all_results)) {
  p <- p + annotate("text", x = i, y = max(all_results$MYC_high[i], all_results$MYC_low[i]) * 1.1, label = sprintf("p = %.4f", all_results$p_value[i]), vjust = -0.5, size = 3)
}

# Display the plot
p + theme_classic()
```


```{r}
individual_results <- data.frame()

# Loop over all genes
for (gene_string in genes_to_analyze) {
  # Get a vector of all ORF IDs that belong to the gene of interest
  orf_ids <- ref_table$orf_id[grepl(gene_string, ref_table$gene_name)]

  # Loop over all ORF IDs
  for (orf_id in orf_ids) {
    # Get gene_name for the current ORF
    gene_name <- ref_table$gene_name[ref_table$orf_id == orf_id]

    # Get the counts for each individual sample, including the sample ID
    cts <- plotCounts(dds = dds, gene = orf_id, intgroup = c("MYC.group", "Sample_ID"), returnData = T, normalized = T)

    # Perform a t-test for this ORF and get the p-value

    # Create a temporary data frame to store the results
    temp_df <- data.frame(gene_name = gene_name,
                          orf_id = orf_id,
                          MYC_status = cts$MYC.group,
                          sampleID = cts$Sample_ID, # Include the sample ID
                          count = cts$count)
    # Append to individual_results
    individual_results <- rbind(individual_results, temp_df)
  }
}

write.csv(individual_results, "riboseq/results/tables/tophits_Psite_counts_long.csv")

```

```{r}
library(dplyr)

# Group by ORF ID and MYC status, and calculate the average count for each group
average_counts <- individual_results %>%
  group_by(orf_id, MYC_status) %>%
  summarise(average_count = mean(count)) %>%
  ungroup()

# Pivot the data to have separate columns for MYC high and MYC low
average_counts_wide <- average_counts %>%
  pivot_wider(names_from = MYC_status, values_from = average_count, names_prefix = "average_")

# View the table with average counts
average_counts_wide
```
```{r}
# Calculate the p-value for each ORF by performing a t-test between MYC high and MYC low
p_values_summary <- individual_results %>%
  group_by(orf_id) %>%
  summarise(p_value = t.test(count[MYC_status == "MYC high"], count[MYC_status == "MYC low"])$p.value) %>%
  ungroup()

# Merge the p-values with the average counts
summary_table <- left_join(average_counts_wide, p_values_summary, by = "orf_id")

# Add gene names to the summary table
summary_table <- left_join(summary_table, distinct(select(individual_results, orf_id, gene_name)), by = "orf_id")

# View the summary table
summary_table

write.csv(summary_table, "riboseq/results/results/tophits_Psite_counts_summary.csv")
```

```{r}
get_ORF_counts <- function(dds, ref_table, genes) {
  # Initialize a data frame to collect results
  results_df <- data.frame()

  # Loop over all genes
  for (gene_string in genes) {
    # Get a vector of all ORF IDs that belong to the gene of interest
    orf_ids <- ref_table$orf_id[grepl(gene_string, ref_table$gene_name)]

    # Loop over all ORF IDs and gather results
    for (orf_id in orf_ids) {
      # Get gene_name for the current ORF
      gene_name <- ref_table$gene_name[ref_table$orf_id == orf_id]

      # Get the counts for each individual sample
      cts <- plotCounts(dds = dds, gene = orf_id, intgroup = c("MYC.group", "Sample_ID"), returnData = T, normalized = T)

      # Create a temporary data frame to store the results
      temp_df <- data.frame(gene_name = gene_name,
                            orf_id = orf_id,
                            MYC_status = cts$MYC.group,
                            sampleID = cts$Sample_ID,
                            count = cts$count)

      # Append to overall results
      results_df <- rbind(results_df, temp_df)
    }
  }

  # Return the results
  return(results_df)
}

# List of genes to analyze
genes_to_analyze <- c("TPT1-AS1", "ZNF525", "CHROMR", "PAX8-AS1", "AC005476.2", "LINC00888", "GARS1-DT", "LERFS")

# Call the function with the list of genes
all_ORF_counts <- get_ORF_counts(dds, ref_table, genes_to_analyze)

write.csv(all_ORF_counts, "riboseq/results/tables/tophits_Psite_counts_long.csv")

```
